move handlers to more global directory
